create and edit working under Manage My Locations
