+++ Spend-It/Controllers/LocationsController.cs
@@ -25,12 +25,13 @@ namespace Spend_It.Controllers
         //GET current signed-in user
         private Task<ApplicationUser> GetCurrentUserAsync() => _userManager.GetUserAsync(HttpContext.User);
 
-        // GET: All Locations
+        // GET: All Locations from every user
         public async Task<IActionResult> Index()
         {
             var applicationDbContext = _context.Locations.Include(l => l.City).Include(l => l.LocationType).Include(l => l.User);
             return View(await applicationDbContext.ToListAsync());
         }
+
 
         // GET: Locations/Details/5
         public async Task<IActionResult> Details(int? id)
@@ -53,132 +54,40 @@ namespace Spend_It.Controllers
             return View(location);
         }
 
-        // GET: Locations/Create
-        public IActionResult Create()
-        {
-            ViewData["CityId"] = new SelectList(_context.Cities, "CityId", "CityName");
-            ViewData["LocationTypeId"] = new SelectList(_context.LocationTypes, "LocationTypeId", "LocationTypeName");
-            //ViewData["UserId"] = new SelectList(_context.ApplicationUsers, "Id", "Id");
-            return View();
-        }
+        //// GET: Locations/Create
+        //public IActionResult Create()
+        //{
+        //    ViewData["CityId"] = new SelectList(_context.Cities, "CityId", "CityName");
+        //    ViewData["LocationTypeId"] = new SelectList(_context.LocationTypes, "LocationTypeId", "LocationTypeName");
+        //    //ViewData["UserId"] = new SelectList(_context.ApplicationUsers, "Id", "Id");
+        //    return View();
+        //}
 
-        // POST: Locations/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("LocationId,DateCreated,Description,LocationName,StreetAddress,UserId,CityId,LocationTypeId")] Location location)
-        {
+        //// POST: Locations/Create
+        //// To protect from overposting attacks, please enable the specific properties you want to bind to, for 
+        //// more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        //[HttpPost]
+        //[ValidateAntiForgeryToken]
+        //public async Task<IActionResult> Create([Bind("LocationId,DateCreated,Description,LocationName,StreetAddress,UserId,CityId,LocationTypeId")] Location location)
+        //{
 
-            ModelState.Remove("User");
-            ModelState.Remove("UserId");
+        //    ModelState.Remove("User");
+        //    ModelState.Remove("UserId");
 
-            if (ModelState.IsValid)
-            {
-                var CurrentUser = await GetCurrentUserAsync();
-                location.UserId = CurrentUser.Id;
-                _context.Add(location);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
-            }
-            ViewData["CityId"] = new SelectList(_context.Cities, "CityId", "CityName", location.CityId);
-            ViewData["LocationTypeId"] = new SelectList(_context.LocationTypes, "LocationTypeId", "LocationTypeName", location.LocationTypeId);
-            ViewData["UserId"] = new SelectList(_context.ApplicationUsers, "Id", "Id", location.UserId);
-            return View(location);
-        }
+        //    if (ModelState.IsValid)
+        //    {
+        //        var CurrentUser = await GetCurrentUserAsync();
+        //        location.UserId = CurrentUser.Id;
+        //        _context.Add(location);
+        //        await _context.SaveChangesAsync();
+        //        return RedirectToAction(nameof(Index));
+        //    }
+        //    ViewData["CityId"] = new SelectList(_context.Cities, "CityId", "CityName", location.CityId);
+        //    ViewData["LocationTypeId"] = new SelectList(_context.LocationTypes, "LocationTypeId", "LocationTypeName", location.LocationTypeId);
+        //    ViewData["UserId"] = new SelectList(_context.ApplicationUsers, "Id", "Id", location.UserId);
+        //    return View(location);
+        //}
 
-        // GET: Locations/Edit/5
-        public async Task<IActionResult> Edit(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var location = await _context.Locations.FindAsync(id);
-            if (location == null)
-            {
-                return NotFound();
-            }
-            ViewData["CityId"] = new SelectList(_context.Cities, "CityId", "CityName", location.CityId);
-            ViewData["LocationTypeId"] = new SelectList(_context.LocationTypes, "LocationTypeId", "LocationTypeName", location.LocationTypeId);
-            ViewData["UserId"] = new SelectList(_context.ApplicationUsers, "Id", "Id", location.UserId);
-            return View(location);
-        }
-
-        // POST: Locations/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("LocationId,DateCreated,Description,LocationName,StreetAddress,UserId,CityId,LocationTypeId")] Location location)
-        {
-            if (id != location.LocationId)
-            {
-                return NotFound();
-            }
-
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(location);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!LocationExists(location.LocationId))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            ViewData["CityId"] = new SelectList(_context.Cities, "CityId", "CityName", location.CityId);
-            ViewData["LocationTypeId"] = new SelectList(_context.LocationTypes, "LocationTypeId", "LocationTypeName", location.LocationTypeId);
-            ViewData["UserId"] = new SelectList(_context.ApplicationUsers, "Id", "Id", location.UserId);
-            return View(location);
-        }
-
-        // GET: Locations/Delete/5
-        public async Task<IActionResult> Delete(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var location = await _context.Locations
-                .Include(l => l.City)
-                .Include(l => l.LocationType)
-                .Include(l => l.User)
-                .FirstOrDefaultAsync(m => m.LocationId == id);
-            if (location == null)
-            {
-                return NotFound();
-            }
-
-            return View(location);
-        }
-
-        // POST: Locations/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
-        {
-            var location = await _context.Locations.FindAsync(id);
-            _context.Locations.Remove(location);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
-        }
-
-        private bool LocationExists(int id)
-        {
-            return _context.Locations.Any(e => e.LocationId == id);
-        }
+      
     }
 }
