+++ Spend-It/Views/MyLocations/Create.cshtml
@@ -1,4 +1,4 @@
-﻿@*@model Spend_It.Models.Location
+﻿@model Spend_It.Models.Location
 
 @{
     ViewData["Title"] = "Create";
@@ -49,4 +49,4 @@
 
 @section Scripts {
     @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
-}*@
+}
