+++ Spend-It/Views/MyLocations/Details.cshtml
@@ -1,0 +1,61 @@
+﻿@model Spend_It.Models.Location
+
+@{
+    ViewData["Title"] = "Details";
+}
+
+<h1>Details</h1>
+
+<div>
+    <h4>Location</h4>
+    <hr />
+    <dl class="row">
+        @*<dt class = "col-sm-2">
+                @Html.DisplayNameFor(model => model.DateCreated)
+            </dt>
+            <dd class = "col-sm-10">
+                @Html.DisplayFor(model => model.DateCreated)
+            </dd>*@
+        <dt class="col-sm-2">
+            @Html.DisplayNameFor(model => model.LocationName)
+        </dt>
+        <dd class="col-sm-10">
+            @Html.DisplayFor(model => model.LocationName)
+        </dd>
+        <dt class="col-sm-2">
+            @Html.DisplayNameFor(model => model.Description)
+        </dt>
+        <dd class="col-sm-10">
+            @Html.DisplayFor(model => model.Description)
+        </dd>
+
+        @*<dt class = "col-sm-2">
+                @Html.DisplayNameFor(model => model.StreetAddress)
+            </dt>
+            <dd class = "col-sm-10">
+                @Html.DisplayFor(model => model.StreetAddress)
+            </dd>
+            <dt class = "col-sm-2">
+                @Html.DisplayNameFor(model => model.User)
+            </dt>
+            <dd class = "col-sm-10">
+                @Html.DisplayFor(model => model.User.Id)
+            </dd>
+            <dt class = "col-sm-2">
+                @Html.DisplayNameFor(model => model.City)
+            </dt>
+            <dd class = "col-sm-10">
+                @Html.DisplayFor(model => model.City.CityName)
+            </dd>
+            <dt class = "col-sm-2">
+                @Html.DisplayNameFor(model => model.LocationType)
+            </dt>
+            <dd class = "col-sm-10">
+                @Html.DisplayFor(model => model.LocationType.LocationTypeName)
+            </dd>*@
+    </dl>
+</div>
+<div>
+    <a asp-action="Edit" asp-route-id="@Model.LocationId">Edit</a> |
+    <a asp-action="Index">Back to List</a>
+</div>
