+++ Spend-It/Views/Shared/_Layout.cshtml
@@ -33,10 +33,10 @@
                             <a class="nav-link text-dark" asp-area="" asp-controller="Locations" asp-action="Index">Home</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Saved" asp-action="Index">My Saved Locations</a>
+                            <a class="nav-link text-dark" asp-area="" asp-controller="SavedLocations" asp-action="Index">My Saved Locations</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Locations" asp-action="Create">Manage My Locations</a>
+                            <a class="nav-link text-dark" asp-area="" asp-controller="MyLocations" asp-action="Index">Manage My Locations</a>
                         </li>
                         <li class="nav-item">
                             <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Privacy">Privacy</a>
